# ClaimsDesk.API/DTOs/v1/PermissionDto.cs
namespace ClaimsDesk.API.DTOs.v1
{
    public class PermissionDto
    {
        public int Id { get; set; }
        public string Name { get; set; } = string.Empty;
        public string Description { get; set; } = string.Empty;
    }
}